+++ server/1ªEvaluacion/bookstore/src/main/java/fpmislata/bookstore/c_domain/_1usecase/admin/author/impl/AuthorFindByIdUseCaseImpl.java
@@ -14,7 +14,7 @@ public class AuthorFindByIdUseCaseImpl implements AuthorFindByIdUseCase {
     private final AuthorService authorService;
 
     @Override
-    public Author execute(Integer id) {
+    public Author execute(Long id) {
         return authorService.findById(id);
     }
 }
