+++ server/1ªEvaluacion/bookstore/src/main/java/fpmislata/bookstore/c_domain/_1usecase/admin/author/interfaces/AuthorCreateUseCase.java
@@ -2,6 +2,6 @@ package fpmislata.bookstore.c_domain._1usecase.admin.author.interfaces;
 
 public interface AuthorCreateUseCase {
 
-    Integer execute(String name, String nationality, String biographyEs, String biographyEn, Integer birthYear,
+    Long execute(String name, String nationality, String biographyEs, String biographyEn, Integer birthYear,
             Integer deathYear);
 }
