+++ server/1ªEvaluacion/bookstore/src/main/java/fpmislata/bookstore/c_domain/_1usecase/admin/author/interfaces/AuthorDeleteUseCase.java
@@ -1,5 +1,5 @@
 package fpmislata.bookstore.c_domain._1usecase.admin.author.interfaces;
 
 public interface AuthorDeleteUseCase {
-    Boolean execute(Integer id);
+    Boolean execute(Long id);
 }
