+++ server/1ªEvaluacion/bookstore/src/main/java/fpmislata/bookstore/c_domain/_1usecase/admin/author/interfaces/AuthorUpdateUseCase.java
@@ -1,6 +1,6 @@
 package fpmislata.bookstore.c_domain._1usecase.admin.author.interfaces;
 
 public interface AuthorUpdateUseCase {
-    Boolean execute(Integer id, String name, String nationality, String biographyEn, String biographyEs,
+    Boolean execute(Long id, String name, String nationality, String biographyEn, String biographyEs,
             Integer birthYear, Integer deathYear);
 }
